finished entry level ticket work

+++ c-beg-to-adv/ticket-work/batch-01/dsp-core-system/src/format_adc_line.c
@@ -4,5 +4,5 @@
 
 int format_adc_line(char* const buf, size_t const buf_size, int const counts, double const volts)
 {
-    return snprintf(buf, buf_size, "counts=%d volts=%d", counts, (int)volts);
+    return snprintf(buf, buf_size, "counts=%d volts=%5.4lf", counts, volts);
 }
